Tools2018CortielWagnerEder/janoliver20_novychen_ntcw_android/dummy/OPR_UE06_MC18_S1810237031/OPR_UE06/src/at/fhhgb/mc/Aufgabe02/ChairMember.java
package at.fhhgb.mc.Aufgabe02;

public class ChairMember extends AbstractMember {

	private int competenceLevel;

	public ChairMember(int competenceLevel, String name) throws ValueException {
		super(name);

		if (competenceLevel < 0 || competenceLevel > 10) {
			throw new ValueException(competenceLevel);
		} else {
			this.competenceLevel = competenceLevel;
		}

	}

	@Override
	public double getIncome() {
		return 100 * competenceLevel;
	}

	@Override
	public double getCosts() {

		return 0.2 * getIncome();
	}

}
